ilk kod gözden geçirme önerileri
